第二次提交

+++ src/main/java/com/example/common/utils/SecurityUtil.java
@@ -27,7 +27,7 @@ public class SecurityUtil {
      * 获取当前用户id
      * @return
      */
-    public static Integer userId(){return  getUserInfo().getUserId(); }
+    public static Long userId(){return  getUserInfo().getUserId(); }
 
     /**
      * 获取当前用户账号
